refactor: create (de)provision pipelines
